First two (and simplest) scorers.
reciprocal_rank_at_recall gives 1 over the rank of the final document
needed to have found documents relevant to all the subtopics

precision_at_recall gives the precision at that rank. we count a
document to be relevant if it provides a new subtopic (and ignore
irrelevant already seen subtopics). if a document provides no new
subtopics, it is not considered relevant.

both of these use binary relevance but are easily updated to make use
of grades.