A new scorer, modified_precision_at_recall.py
This score provides the precision at full recall where in this case a
document is counted towards precision in terms of the fraction of new
subtopics (out of total subtopics) it provides. This may or may not be
a good idea.

For now, this just uses binary relevance, but it is easily upgraded.